moved DialogHolder handling to challenge handler
